Added airbus optical adaptor
