+++ airbus/airbus_optical_adaptor/Dockerfile
@@ -7,8 +7,9 @@ RUN python3 -m pip install --upgrade pip
 
 WORKDIR /app
 
-COPY requirements.txt /app
+COPY ../requirements.txt /app
 RUN python -m pip install -r requirements.txt 
 
-COPY . .
-ENTRYPOINT ["python", "-m", "airbus_sar_adaptor"]
+COPY ../common .
+COPY ../airbus_optical_adaptor .
+ENTRYPOINT ["python", "-m", "airbus_optical_adaptor"]
