+++ airbus/common/auth_utils.py
@@ -1,0 +1,46 @@
+import base64
+
+import requests
+from kubernetes import client, config
+
+
+def get_api_key_from_secret(
+    secret_name: str, secret_key: str, namespace: str = "ws-airbus"
+) -> str:
+    """Retrieve an API key from a Kubernetes secret"""
+    # Create a Kubernetes API client
+    config.load_incluster_config()
+    v1 = client.CoreV1Api()
+
+    # Retrieve and decode the secret
+    secret = v1.read_namespaced_secret(secret_name, namespace)
+    api_key_base64 = secret.data[secret_key]
+    api_key = base64.b64decode(api_key_base64).decode("utf-8")
+
+    return api_key
+
+
+def generate_access_token(env: str = "prod") -> str:
+    """Generate an access token for the Airbus OneAtlas API"""
+    api_key = get_api_key_from_secret("api-keys", "airbus-key")
+    if not api_key:
+        raise ValueError("API key not found in secret")
+
+    if env == "prod":
+        url = "https://authenticate.foundation.api.oneatlas.airbus.com/auth/realms/IDP/protocol/openid-connect/token"
+    else:
+        url = "https://authenticate-int.idp.private.geoapi-airbusds.com/auth/realms/IDP/protocol/openid-connect/token"
+
+    headers = {
+        "Content-Type": "application/x-www-form-urlencoded",
+    }
+
+    data = [
+        ("apikey", api_key),
+        ("grant_type", "api_key"),
+        ("client_id", "IDP"),
+    ]
+
+    response = requests.post(url, headers=headers, data=data)
+
+    return response.json()["access_token"]
