+++ airbus/common/stac_utils.py
@@ -1,0 +1,147 @@
+import json
+import logging
+import mimetypes
+from enum import Enum
+from common.pulsar_utils import send_pulsar_message
+from common.s3_utils import retrieve_stac_item, upload_stac_item, list_objects_in_folder
+
+
+class OrderStatus(Enum):
+    ORDERABLE = "orderable"
+    ORDERED = "ordered"
+    PENDING = "pending"
+    SHIPPING = "shipping"
+    SUCCEEDED = "succeeded"
+    FAILED = "failed"
+    CANCELED = "canceled"
+
+
+def write_stac_item_and_catalog(stac_item: dict, stac_item_filename: str, item_id: str):
+    """Creates local catalog containing final STAC item to be used as a record for the order"""
+    # Rewrite STAC links to point to local files only
+    stac_item["links"] = [
+        {"rel": "self", "href": stac_item_filename, "type": "application/json"},
+        {"rel": "parent", "href": "catalog.json", "type": "application/json"},
+    ]
+
+    # Write the STAC item to a file
+    with open(stac_item_filename, "w") as f:
+        json.dump(stac_item, f, indent=2)
+    logging.info(f"Created STAC item '{stac_item_filename}' locally.")
+    logging.debug(f"STAC item: {stac_item}")
+
+    # If not item_id, the order has failed
+    if not item_id:
+        item_id = "Failed"
+
+    # Create containing STAC catalog
+    stac_catalog = {
+        "stac_version": "1.0.0",
+        "id": "catalog",
+        "type": "Catalog",
+        "description": f"Root catalog for order {stac_item_filename}-{item_id}",
+        "links": [
+            {"rel": "self", "href": "catalog.json", "type": "application/json"},
+            {"rel": "item", "href": stac_item_filename, "type": "application/json"},
+        ],
+    }
+
+    # Write the STAC catalog to a file
+    with open("catalog.json", "w") as f:
+        json.dump(stac_catalog, f, indent=2)
+    logging.info("Created STAC catalog catalog.json locally.")
+    logging.debug(f"STAC catalog: {stac_catalog}")
+
+
+def update_stac_order_status(stac_item: dict, item_id: str, order_status: str):
+    """Update the STAC item with the order status using the STAC Order extension"""
+    # Update or add fields relating to the order
+    if "properties" not in stac_item:
+        stac_item["properties"] = {}
+
+    if item_id is not None:
+        stac_item["properties"]["order.id"] = item_id
+    stac_item["properties"]["order.status"] = order_status
+
+    # Update or add the STAC extension if not already present
+    order_extension_url = "https://stac-extensions.github.io/order/v1.1.0/schema.json"
+    if "stac_extensions" not in stac_item:
+        stac_item["stac_extensions"] = []
+
+    if order_extension_url not in stac_item["stac_extensions"]:
+        stac_item["stac_extensions"].append(order_extension_url)
+
+
+def get_acquisition_id_from_stac(stac_item: dict, key: str) -> str:
+    """Extract the acquisition ID from a STAC item"""
+    acquisition_id = stac_item.get("properties", {}).get("acquisition_id")
+    if not acquisition_id:
+        raise ValueError(f"Acquisition ID not found in STAC item '{key}'.")
+    return acquisition_id
+
+
+def get_key_from_stac(stac_item: dict, key: str):
+    """Extract a nested key from a STAC item. Key given as a dot-separated string."""
+    parts = key.split(".")
+    value = stac_item
+    for part in parts:
+        value = value.get(part)
+        if value is None:
+            logging.error(f"{part} not found in STAC item.")
+            return None
+    return value
+
+
+def update_stac_item_failure(bucket: str, key: str, item_id: str):
+    """Update the STAC item with the failure order status"""
+    stac_item = retrieve_stac_item(bucket, key)
+
+    # Mark the order as failed and upload the updated STAC item
+    update_stac_order_status(stac_item, item_id, OrderStatus.FAILED.value)
+    upload_stac_item(bucket, key, stac_item)
+    send_pulsar_message(bucket, key)
+
+    # Create local record of attempted order, to be used as the workflow output
+    write_stac_item_and_catalog(stac_item, key.split("/")[-1], item_id)
+
+
+def update_stac_item_success(
+    bucket: str, key: str, parent_folder: str, item_id: str, workspaces_domain: str
+):
+    """Update the STAC item with the assets and success order status"""
+    stac_item = retrieve_stac_item(bucket, key)
+    # List files and folders in the specified folder
+    folder_prefix = f"{parent_folder}/{item_id}/"
+    folder_objects = list_objects_in_folder(bucket, folder_prefix)
+
+    # Add all listed objects as assets to the STAC item
+    if "Contents" in folder_objects:
+        for obj in folder_objects["Contents"]:
+            file_key = obj["Key"]
+
+            # Skip if the file_key is a folder
+            if file_key.endswith("/"):
+                continue
+
+            asset_name = file_key.split("/")[-1]
+
+            # Determine the MIME type of the file
+            mime_type, _ = mimetypes.guess_type(file_key)
+            if mime_type is None:
+                mime_type = "application/octet-stream"  # Default MIME type
+
+            # Add asset link to the file
+            parts = file_key.split("/", 1)
+            workspace = parts[0]
+            file_subpath = parts[1]
+            stac_item["assets"][asset_name] = {
+                "href": f"https://{workspace}.{workspaces_domain}/files/{bucket}/{file_subpath}",
+                "type": mime_type,
+            }
+    # Mark the order as succeeded and upload the updated STAC item
+    update_stac_order_status(stac_item, item_id, OrderStatus.SUCCEEDED.value)
+    upload_stac_item(bucket, key, stac_item)
+    send_pulsar_message(bucket, key)
+
+    # Create local record of the order, to be used as the workflow output
+    write_stac_item_and_catalog(stac_item, key.split("/")[-1], item_id)
